Added local login authentication for lab 12

+++ routes/tasks.js
@@ -2,9 +2,27 @@ var express = require('express');
 var router = express.Router();
 var Task = require('../models/task');
 
-/* GET home page. */
+/* need to make sure user is logged in before doing tasks
+* create a middleware function to check if user is logged in
+* redirect to authentication page if not */
+
+function isLoggedIn(req, res, next) {
+    console.log('user is auth', req.user)
+    if (req.isAuthenticated()) {
+        res.locals.username = req.user.local.username;
+        next();
+    } else {
+        res.redirect('/auth')
+    }
+}
+
+/* this requires all routes in this file to use
+* isLoggedIn middleware, but we dont need to specify it */
+router.use(isLoggedIn);
+
+/* GET home page, a list of incomplete tasks for the current user */
 router.get('/', function(req, res, next) {
-  Task.find( {completed: false} )
+  Task.find({_creator:req.user, completed: false})
       .then( (docs) => {
         res.render('index', {title: 'Incomplete tasks', tasks: docs} );
       })
@@ -18,9 +36,13 @@ router.get('/', function(req, res, next) {
 router.post('/add', function(req, res, next){
 
   // check if something was entered in the text input
-  if (req.body.text) {
+  if (!req.body || !req.body.text) {
+      req.flash('error', 'Please enter some text');
+      res.redirect('/');
+  }
+  else {
       // create new Task
-      var t = new Task({text: req.body.text, completed: false})
+      var t = new Task({_creator: req.user, text: req.body.text, completed: false})
       // save the task, and redirect to home page if successful
       t.save().then((newTask) => {
           console.log('The new task created is ', newTask); // just for debugging
@@ -29,27 +51,22 @@ router.post('/add', function(req, res, next){
           next(err); // forward error to the error handlers
       });
   }
-  else {
-    req.flash('error', 'Please enter a task.') // Flash an error message
-    res.redirect('/'); // else, ignore and redirect to home page. We'll improve this later
-  }
 
 });
 
-/* POST to mark a task as done */
+/* POST to mark a task as done. Task _id should be provided as req.body parameter*/
 
 router.post('/done', function(req, res, next){
 
-  Task.findByIdAndUpdate( req.body._id, {completed: true})
+  Task.findByIdAndUpdate( {_id: req.body._id, _creator:req.user.id}, {completed: true})
       .then( (originalTask) => {
         // originalTask only has a value if a document with this _id was found
-        if (originalTask) {
-          req.flash('info', originalTask.text + ' marked as done!');
-          res.redirect('/'); // redirect to list of tasks
-        }  else {
-          var err = new Error('Not Found'); // report Task not found with 404 status
-          err.status = 404;
-          next(err);
+        if (!originalTask) {
+            res.status(403).send('This is not your task!');
+        }
+        else {
+            req.flash('info', originalTask.text + ' marked as done!');
+            res.redirect('/'); // redirect to list of tasks
         }
       })
       .catch( (err) => {
@@ -60,7 +77,7 @@ router.post('/done', function(req, res, next){
 /* GET completed tasks */
 router.get('/completed', function(req, res, next){
 
-  Task.find( {completed:true} )
+  Task.find( {creator: req.user._id, completed:true} )
       .then( (docs) => {
       res.render('completed_tasks', {tasks:docs});
       }).catch( (err) => {
@@ -69,19 +86,18 @@ router.get('/completed', function(req, res, next){
 
 });
 
-/* POST to delete a task */
+/* POST to delete a task. Task _id is in req.body */
 
 router.post('/delete', function(req, res, next){
 
-  Task.findByIdAndRemove(req.body._id)
+  Task.findByIdAndRemove({_id: req.body._id, _creator:req.user.id}, {completed:true})
       .then( (deletedTask) => {
-        if (deletedTask) {
-          req.flash('info', 'Task deleted.');
-          res.redirect('/');
-        } else {
-          var error = new Error('Task Not Found')
-          error.status = 404;
-          next(error);
+        if (!deletedTask) {
+            res.status(403).send('This is not your task!');
+        }
+        else {
+            req.flash('info', 'Task deleted.');
+            res.redirect('/');
         }
       })
       .catch( (err) => {
@@ -93,7 +109,7 @@ router.post('/delete', function(req, res, next){
 
 router.post('/alldone', function(req, res, next){
 
-  Task.updateMany({completed:false}, {completed: true})
+  Task.updateMany({_creator:req.user, completed:false}, {completed: true}, {multi:true})
       .then( () => {
         req.flash('info', 'All tasks are done!');
         res.redirect('/'); // if preferred, redirect to /completed
@@ -108,11 +124,16 @@ router.get('/task/:_id', function(req, res, next){
 
   Task.findById(req.params._id)
       .then( (doc) => {
-        if (doc) {
-          res.render('task', {task:doc});
+        if (!doc) {
+            res.status(404).send('Task not found.');
         }
+        // verify this task was created by the currently logged in user
+        else if (!doc._creator.equals(req.user._id)) {
+            res.status(403).send('This is not your task!');
+        }
+
         else {
-          next(); // to the 404 error handler
+            res.render('task', {task:doc})
         }
       })
       .catch( (err) => {
@@ -123,7 +144,7 @@ router.get('/task/:_id', function(req, res, next){
 /* POST delete all completed tasks */
 router.post('/deleteDone', function (req, res, next) {
 
-        Task.deleteMany({completed: true})
+        Task.deleteMany({_creator:req.user}, {completed: true})
             .then( (deletedTask) => {
                 if (deletedTask) {
                     res.redirect('/');
